+++ django-rgd-3d/rgd_3d/admin/__init__.py
@@ -1,1 +1,1 @@
-from .point_cloud import *  # noqa
+from .mesh import *  # noqa
