+++ django-rgd-3d/rgd_3d/models/mesh.py
